+++ public/inject.js
@@ -11,5 +11,10 @@ function injectScript(file_path, tag) {
   script.setAttribute('type', 'text/javascript');
   script.setAttribute('src', file_path);
   node.appendChild(script);
+
+  var extensionIdScript = document.createElement('script');
+  extensionIdScript.setAttribute("type", "application/javascript");
+  extensionIdScript.textContent = 'window.extensionId = \'' + chrome.runtime.id + '\';';
+  node.appendChild(extensionIdScript);
 }
 injectScript(chrome.extension.getURL('content.js'), 'body');